+++ packages/zowe-explorer/src/trees/job/JobInit.ts
@@ -60,9 +60,7 @@ export class JobInit {
                 }
             })
         );
-        context.subscriptions.push(
-            vscode.commands.registerCommand("zowe.jobs.refreshJobsServer", async (job) => JobActions.refreshJob(job, jobsProvider))
-        );
+        context.subscriptions.push(vscode.commands.registerCommand("zowe.jobs.refreshJobsServer", (job) => JobActions.refreshJob(job, jobsProvider)));
         context.subscriptions.push(
             vscode.commands.registerCommand("zowe.jobs.refreshAllJobs", async () => {
                 await SharedActions.refreshAll(jobsProvider);